Hestia.DynamicAnisotropic

Hestia.DynamicIsotropic

Hestia.HeatCuboid

Hestia.HeatPlate

Hestia.HeatPlate

Hestia.HeatRod

Hestia.IOSetup

Hestia.RadialCharacteristics

Hestia.RadialCharacteristics

Hestia.RadialCharacteristics

Hestia.RadialCharacteristics

Hestia.RadialCharacteristics

Hestia.StaticAnisotropic

Hestia.Boundary

Hestia.characterize

Hestia.characterize

Hestia.getCharacteristics

Hestia.getdiffusivity

Hestia.getindices

Hestia.getindices

Hestia.measure

Hestia.measure

Properties

Hestia.DynamicAnisotropic — Method

createDynamicAnisotropic(conductivityx :: Vector{T}, conductivityy :: Vector{T}, conductivity_z :: Vector{T}, density :: Vector{T}, capacity :: Vector{T}) where T <: Real

Returns a DynamicAnisotropic

Hestia.DynamicIsotropic — Type

DynamicIsotropic

Temperature-dependent isotropic properties

A record for static isotropic properties.

### Elements  
`λᵢ` : thermal conductivity coefficients 

`ρᵢ` : volumetric mass density coefficients

`cᵢ` : specific heat capacity coefficients

Hestia.StaticAnisotropic — Type

Static anisotropic properties λ = diag(λx, λy, λz) = [λx 0 0; 0 λy 0 ; 0 0 λz]

Hestia.getdiffusivity — Method

getdiffusivity(prop :: StaticIsotropic)

Returns the diffusivity of a StaticIsotropic

α = λ/(ρ ⋅ c)

Geometry

Hestia.HeatCuboid — Type

HeatCuboid <: AbstractCubicObject

Model of a two dimensional plate for heat conduction

Elements

dimension : tuple of length, width and heigth of the plate: (length, width, height)

sampling : tuple of spatial discretization: (Δx, Δy, Δz)

heatcells : number of heatcells in total

Hestia.HeatPlate — Type

HeatPlate <: AbstractCubicObject

Model of a two dimensional plate for heat conduction

Elements

dimension : tuple of length and width of the plate: (length, width)

sampling : tuple of spatial discretization: (Δx, Δy)

heatcells : Number of heatcells per direction: {Nx, Ny}

Hestia.HeatPlate — Method

HeatPlate(plate_length :: Real, plate_width ::  Real, Nx :: Integer, Ny :: Integer, heatcells :: Array{S,1} where S <: Real)

Returns a HeatPlate model

Hestia.HeatRod — Method

HeatRod(rod_length :: Real,  heatcells :: Integer )

Returns a HeatRod model

Hestia.getindices — Method

getindices( heatPlate :: HeatPlate ; cellPosition::Symbol = :complete )

Returns all indices for a certain cellPosition

Valid CellPositions

:complete

:center

:west

:east

:south

:north

Hestia.getindices — Method

getindices( heatrod :: HeatRod ; cellPosition::Symbol = :complete )

Returns all indices for a certain cellPosition

Valid CellPositions

:complete

:center

:west

:east

Boundary

Hestia.Boundary — Method

Boundary(geometry :: AbstractGeometricalObject)

Initialize the boundary sides for a geometry.

Configuration

Hestia.RadialCharacteristics — Type

RadialCharacteristics <: AbstractStaticCharacteristics

Stores the values for the calculation of $m ~ exp( -||M (x - xₛ)||^{2ν} )$

scaling : m ∈ [0, 1]

power : ν ∈ [0, ∞)

center : xₛ ∈ R^{3}

curvature : M ∈ R^{3 x 3} for the planar boundaries

Hestia.RadialCharacteristics — Method

RadialCharacteristics( scale :: Real , power :: Integer, curvature :: Matrix{<: Real} )

The center is set to the origin (0,0,0).

Hestia.RadialCharacteristics — Method

RadialCharacteristics( scale :: Real , power :: Integer, curvature :: Real)

The center is set to the origin (0,0,0).

Hestia.RadialCharacteristics — Method

RadialCharacteristics( scale :: Real , power :: Integer, central_point :: Tuple{Real,Real,Real}, curvature :: Real)

Here the variable curvature is multiplied with the identity matrix to gain matrix M.

Returns a RadialCharacteristics.

Hestia.RadialCharacteristics — Method

RadialCharacteristics()

Initializes and returns a basic RadialCharacteristics

Hestia.characterize — Method

characterize(start :: Real, stop :: Real, step :: Real, config :: RadialCharacteristics; dim = 1 :: Integer )

x₁: dim = 1

x₂: dim = 2

x₃: dim = 3

Hestia.characterize — Method

characterize(start :: Real, stop :: Real, step :: Real, config :: RadialCharacteristics; dim = 1 :: Integer )

x₁ ∪ x₂ dim = {1,2}

x₁ ∪ x₃: dim = {1,3}

x₂ ∪ x₃: dim = {2,3}

IOSetup

Hestia.IOSetup — Method

IOSetup(geometry :: AbstractGeometricalObject)

Initialize IOSetup as basic element for actuation and sensing.

Hestia.getCharacteristics — Method

getCharacteristics(iosetup :: IOSetup, orientation :: Symbol)

Returns the characteristics and the identifier of each actuator or sensor for a certain IOSetup and boundary side.

Hestia.measure — Method

measure(temperatures :: Matrix{<:Real}, character :: Vector{<:Real})

Returns the weighted arithmetic mean of the measurement.

Hestia.measure — Method

measure(temperatures :: Vector{<:Real}, character :: Vector{<:Real})

Returns the weighted arithmetic mean of the measurement.