Boundary conditions

The temperature $\theta$ inside the geometrical object is driven by

the material properties and the initial temperature and

the heat flux along the geometrical boundary sides.

We denote the heat flux

inwards the object as induced heat flux $\Phi_{in}$ and

outwards the object as emitted heat flux $\Phi_{out}$ or Emission.

Mathematically spoken, the situation on the boundary sides are described here by Neumann or Robin boundary conditions. On boundary sides where actuators (e.g. heating elements) are assumed, we have

\[\left[\lambda(\theta) \nabla \theta(t,x) \right] ~ \vec{n} = \Phi_{in}(t,x) + \Phi_{out}(t,x)\]

with outer normal vector $\vec{n}$.

Emitted heat flux / Emissions

The emissions are described by

linear heat transfer / convection

\[-h (\theta - \theta_{amb})\]

and nonlinear heat radiation

\[- \epsilon ~ \sigma (\theta - \theta_{amb})\]

with ambient temperature (temperature of the object's surounding) $\theta_{amb}$, heat transfer coefficient $h>0$, emissivity $\epsilon \in [0,1]$ and Stefan-Boltzmann constant $\sigma$.

Heat transfer coefficient $h$ and emissivity $\epsilon$ can be defined for each boundary side seperately. Internally, the multiplication of emissivity $\epsilon$ and Stefan-Boltzmann constant $\sigma$ is saved as radiation coefficient $k = \epsilon ~ \sigma$.

Hestia.Emission — Type

Emission  <: AbstractEmission

Type Emission contains the coefficients for heat transfer (convection) h, heat radiation k, and the ambient temperature θamb to model linear or nonlinear Stefan-Boltzmann boundary conditions

\[\Phi = -h ~ (\theta - \theta_{amb}) -k ~ (\theta^4 - \theta_{amb}^4)\]

Constructor Emission(h,ϵ,θamb) expects emissivity ϵ which must be in interval [0,1]. The heat radiation coefficient is calculated internally as k=ϵ⋅σ using the Stefan-Boltzmann constant: σ = 5.6703744191844294e-8.

Elements

h : heat transfer coefficient

k : heat radiation coefficient

θamb : ambient temperature

Hestia.emit — Function

emit(temperature :: Real, emission :: Emission)

Calculates the right-hand side of the boundary conditions for a given Emission.

\[\Phi = -h ~ (\theta - \theta_{amb}) -k ~ (\theta^4 - \theta_{amb}^4)\]

Hestia.emit! — Function

emit!(flux :: Vector{<:Real}, temperature :: Vector{<:Real}, emission :: Emission)

Calculates the right-hand side of the natural Robin boundary along a boundary for a given Emission.

Note: in-place operation - results are saved in array flux.

Induced heat flux

The induced heat flux is generated by $N_{u}$ actuators that can be placed directly (e.g. heating elements) or indirectly (e.g. laser beam) on each boundary of the geometrical object. All actuators are assumed to have a specific spatial characterization $b_{n}(x)$ reaching from 0 to 1 like a scaling from zero to 100%. The control signal $u_{n}(t)$ defines the power intensity of the n-th actuator. So, the induced heat flux is defined by

\[\Phi_{in}(t,x) = \sum_{n=1}^{N_{u}} b_{n}(x) ~ u_{n}(t).\]