Hestia.DynamicIsoProperty

Hestia.Emission

Hestia.HeatCuboid

Hestia.HeatPlate

Hestia.HeatPlate

Hestia.HeatRod

Hestia.HyperSegment

Hestia.MixedSegment

Hestia.RadialConfiguration

Hestia.SimpleSegment

Hestia.StaticIsoProperty

Hestia.characterize

Hestia.characterize

Hestia.createDynamicIsoProperty

Hestia.createEmission

Hestia.createSimpleSegment

Hestia.createStaticIsoProperty

Hestia.emit

Hestia.emit!

Hestia.getdiffusivity

Hestia.getindices

Hestia.getindices

Hestia.initConfiguration

Hestia.setConfiguration

Hestia.setConfiguration

Hestia.setConfiguration

Hestia.setConfiguration

Properties

Hestia.DynamicIsoProperty — Type

DynamicIsoProperty

Temperature-dependent isotropic properties

A record for static isotropic properties.

### Elements  
`λᵢ` : thermal conductivity coefficients 

`ρᵢ` : volumetric mass density coefficients

`cᵢ` : specific heat capacity coefficients

Hestia.StaticIsoProperty — Type

StaticIsoProperty

Temperature-independent isotropic properties

A record for static isotropic properties.

Elements

λ : thermal conductivity

ρ : volumetric mass density

c : specific heat capacity

Hestia.createDynamicIsoProperty — Method

createDynamicIsoProperty(conductivity :: Array{T,1}, density :: Array{T,1}, capacity :: Array{T,1})  where T <: Real

Returns a DynamicIsoProperty

Hestia.createStaticIsoProperty — Method

createStaticIsoProperty(conductivity :: Real, density :: Real, capacity :: Real)

Returns a StaticIsoProperty

Hestia.getdiffusivity — Method

getdiffusivity(prop :: StaticIsoProperty)

Returns the diffusivity of a StaticIsoProperty

α = λ/(ρ ⋅ c)

Geometry

Hestia.HeatCuboid — Type

HeatCuboid <: AbstractCubicObject

Model of a two dimensional plate for heat conduction

Elements

dimension : tuple of length, width and heigth of the plate: (length, width, height)

sampling : tuple of spatial discretization: (Δx, Δy, Δz)

heatcells : number of heatcells in total

segmentation : underlying segmentation, e.g. SimpleSegment or MixedSegment

Hestia.HeatPlate — Type

HeatPlate <: AbstractCubicObject

Model of a two dimensional plate for heat conduction

Elements

dimension : tuple of length and width of the plate: (length, width)

sampling : tuple of spatial discretization: (Δx, Δy)

heatcells : Number of heatcells per direction: {Nx, Ny}

segmentation : underlying segmentation, e.g. SimpleSegment or MixedSegment

Hestia.HeatPlate — Method

HeatPlate(plate_length :: Real, plate_width ::  Real, Nx :: Integer, Ny :: Integer, heatcells :: Array{S,1} where S <: Real, segmentation :: T where T <: AbstractSegmentation)

Returns a HeatPlate model

Hestia.HeatRod — Method

HeatRod(rod_length :: Real,  heatcells :: Integer, segmentation :: T where T <: AbstractSegmentation )

Returns a HeatRod model

Hestia.getindices — Method

getindices( heatPlate :: HeatPlate ; cellPosition::Symbol = :complete )

Returns all indices for a certain cellPosition

Valid CellPositions

:complete

:center

:west

:east

:south

:north

Hestia.getindices — Method

getindices( heatrod :: HeatRod ; cellPosition::Symbol = :complete )

Returns all indices for a certain cellPosition

Valid CellPositions

:complete

:center

:west

:east

Segmentation

Hestia.HyperSegment — Type

HyperSegment <: AbstractSegmentation

A meta container segment to store all segments

Hestia.MixedSegment — Type

MixedSegment <: AbstractSegmentation

A container that identifies each element of a list of cells with its own specific thermal property.

Hestia.SimpleSegment — Type

SimpleSegment

A container that identifies a list of cells with its thermal property.

Hestia.createSimpleSegment — Method

createSimpleSegment( property :: AbstractIsotropicProperty, cellindices :: Array{S2, 1} where S2 <: Integer)

Returns a SimpleSegment

Boundary

Emission

Hestia.Emission — Type

Emission  <: AbstractEmission

Returns the right-hand side of a natural Robin boundary condition including heat conduction and emission to the environment.

Elements

conduction : heat conduction coefficient

radiation : heat radiation coefficient

θamb : ambient temperature

Hestia.createEmission — Method

createEmission(conduction :: Real, emissivity :: Real, θamb :: Real)

Returns an Emission for a given conductivity, emissivity and ambient temperature θamb

Information

This constant is used:

Stefan-Boltzmann constant: σ = 5.6703744191844294e-8

Hestia.emit! — Method

emit!(flux :: Array{S1,1} where S1 <: Real, temperature :: Array{S2,1} where S2 <: Real, emission :: Emission)

Calculates the right-hand side of the natural Robin boundary along a boundary for a given Emission.

Note: in-place operation - results are saved in array flux.

Hestia.emit — Method

emit!(flux :: Array{S1,1} where S1 <: Real, temperature :: Array{S2,1} where S2 <: Real, emission :: Emission)

Calculates the right-hand side of the natural Robin boundary along a boundary for a given Emission.

Note: in-place operation - results are saved in array flux.

Configuration

Hestia.RadialConfiguration — Type

RadialConfiguration <: AbstractStaticConfiguration

Stores the values for the calculation of $m ~ exp( -||M (x - xₛ)||^{2ν} )$

scaling : m ∈ [0, 1]

power : ν ∈ [0, ∞)

center : xₛ ∈ R^{3}

curvature : M ∈ R^{3 x 3} for the planar boundaries

Hestia.characterize — Method

characterize(start :: Real, stop :: Real, step :: Real, config :: RadialConfiguration; dim = 1 :: Integer )

x₁: dim = 1

x₂: dim = 2

x₃: dim = 3

Hestia.characterize — Method

characterize(start :: Real, stop :: Real, step :: Real, config :: RadialConfiguration; dim = 1 :: Integer )

x₁ ∪ x₂ dim = {1,2}

x₁ ∪ x₃: dim = {1,3}

x₂ ∪ x₃: dim = {2,3}

Hestia.initConfiguration — Method

initConfiguration()

Initializes and returns a basic RadialConfiguration

Hestia.setConfiguration — Method

setConfiguration( scale :: Real , power :: Integer; curvature = 1.0 :: Real)

The center is set to the origin (0,0,0).

Hestia.setConfiguration — Method

setConfiguration( scale :: Real , power :: Integer, curvature :: Real)

The center is set to the origin (0,0,0).

Hestia.setConfiguration — Method

setConfiguration( scale :: Real , power :: Integer, central_point :: Tuple{Real,Real,Real}, curvature :: Array{T,2} where T <: Real)

Returns a RadialConfiguration

Hestia.setConfiguration — Method

setConfiguration( scale :: Real , power :: Integer, central_point :: Tuple{Real,Real,Real}, curvature :: Real)

Here the variable curvature is multiplied with the identity matrix to gain matrix M.

Returns a RadialConfiguration